Improved email verification experience
