+++ app/Http/Controllers/UserController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Illuminate\Auth\Events\PasswordReset;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Password;
+use Illuminate\Support\Str;
 use function Laravel\Prompts\alert;
 
 class UserController extends Controller
@@ -26,12 +29,6 @@ class UserController extends Controller
         if ($email != $conf_email || $password != $conf_password) {
             return redirect("signup")->with("status", "Something went wrong...\nPlease check details and try again");
         } else {
-//            DB::table("users")->insert([
-//                "name" => $username,
-//                "email" => $email,
-//                "password" => $password,
-//                "phone_number" => $phone_num,
-//            ]);
             $user = new User();
             $user->password = Hash::make($password);
             $user->email = $email;
@@ -39,7 +36,6 @@ class UserController extends Controller
             $user->phone_number = $phone_num;
             $user->save();
             event(new Registered($user));
-//            auth()->login($user);
             return redirect("verify-email");
         }
     }
@@ -66,5 +62,42 @@ class UserController extends Controller
         $request->session()->regenerateToken();
 
         return redirect("/");
+    }
+
+    public function send_reset_email(Request $request) {
+        $request->validate(['email' => 'required|email']);
+
+        $status = Password::sendResetLink(
+            $request->only('email')
+        );
+
+        return $status === Password::RESET_LINK_SENT
+            ? back()->with(['status' => __($status), 'success' => 't'])
+            : back()->withErrors(['email' => __($status)]);
+    }
+
+    public function reset_password(Request $request) {
+        $request->validate([
+            'token' => 'required',
+            'email' => 'required|email',
+            'password' => 'required|min:8|confirmed',
+        ]);
+
+        $status = Password::reset(
+            $request->only('email', 'password', 'password_confirmation', 'token'),
+            function (User $user, string $password) {
+                $user->forceFill([
+                    'password' => Hash::make($password)
+                ])->setRememberToken(Str::random(60));
+
+                $user->save();
+
+                event(new PasswordReset($user));
+            }
+        );
+
+        return $status === Password::PASSWORD_RESET
+            ? back()->with(['status' => __($status), 'success' => 't'])
+            : back()->withErrors(['email' => [__($status)]]);
     }
 }
